requirements.txt
gradio
faiss-cpu
sentence-transformers
pandas
numpy
langchain
langchain-google-genai
